Moved: navbar and footer to shared
