+++ src/features/boolean-algebra/StandardForms.jsx
@@ -5,8 +5,8 @@ import ControlPanel from '../../shared/components/ControlPanel';
 import ControlGroup from '../../shared/components/ControlGroup';
 import CircuitModal from '../../shared/components/CircuitModal';
 import { parseSOP, generateTruthTable } from '../../shared/utils/boolMath';
-import { Navbar } from '../home/Navbar';
-import Footer from '../home/Footer';
+import { Navbar } from '../../shared/components/Navbar';
+import Footer from '../../shared/components/Footer';
 import { useTheme } from '../../shared/context/ThemeContext';
 
 const toPOS = (variables, expression) => {
