+++ src/features/boolforge/Boolforge.jsx
@@ -4,8 +4,8 @@ import { TruthTableGenerator } from "../../shared/components/TruthTable";
 import { SaveAndLoad } from "../../shared/components/SaveAndLoad";
 import { parseExpressionToCircuit } from "../../shared/utils/expressionParser";
 import RelatedSeoLinks from "../../shared/seo/RelatedSeoLinks";
-import { Navbar } from "../home/Navbar";
-import Footer from "../home/Footer";
+import { Navbar } from "../../shared/components/Navbar";
+import Footer from "../../shared/components/Footer";
 import { useTheme } from "../../shared/context/ThemeContext";
 import { getCircuitHint } from "../../shared/services/circuitMindService";
 import { generateAiCircuit } from "../../shared/services/aiService";
