+++ src/features/coal/InstructionLaboratoryPage.jsx
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
-import { Navbar } from "../home/Navbar";
-import Footer from "../home/Footer";
+import { Navbar } from "../../shared/components/Navbar";
+import Footer from "../../shared/components/Footer";
 import { useTheme } from "../../shared/context/ThemeContext";
 import "./InstructionLaboratoryPage.css";
