+++ src/features/sequential-circuits/TimeDiagrams.jsx
@@ -2,8 +2,8 @@ import React, { useState, useEffect, useMemo } from 'react';
 import ToolLayout from '../../shared/components/ToolLayout';
 import ExplanationBlock from '../../shared/components/ExplanationBlock';
 import CircuitModal from '../../shared/components/CircuitModal';
-import { Navbar } from '../home/Navbar';
-import Footer from '../home/Footer';
+import { Navbar } from '../../shared/components/Navbar';
+import Footer from '../../shared/components/Footer';
 import { useTheme } from '../../shared/context/ThemeContext';
 
 const TimeDiagrams = () => {
