+++ src/shared/components/Footer.jsx
